Enhanced Hand Detection and Improved Model Accuracy
Added logic to ensure predictions are only made when hands are detected on screen, enhancing the accuracy and efficiency of real-time ASL detection.
Increased the dataset size for model training, leading to improved prediction accuracy.
Refactored the display logic to replace the detected actions on-screen instead of appending, preventing text overflow.